Prevent FPEs in cooling sims with external radiation fields
Including external radiation fields (e.g., the CMB) in cooling sims
may lead to FPEs.  The cumulative radiation field intensity involves
multiplication by a factor that may be near the maximum value a float
can hold.  This is not a problem with diffuse radiation.  However, when
the CMB is included, the product exceeds that limit and an FPE occurs.

The adopted bugfix hides in a new class the details of how the
integrated flux is accumulated, while preserving how the flux vectors
are accessed.  The latter is mainly useful for the instantaneous flux.

As an added feature, the time-weighted mean flux is also computed.

Reported-by: M. Chatzikos
https://groups.google.com/g/cloudy-dev/c/GSWoDN_hxqc