Add columns for advective terms to save cool/heat
The output of the save cooling/heating commands treats advective heating
and cooling terms ambiguously.  It excludes them from the total when
reporting fractional contributions, but reports them as contributing
agents.  As a result, advective terms with fractions >>100% are commonly
encountered in cooling simulations.

Add columns to report these contributions, but exclude them from the
list of cooling/heating agents.

https://groups.google.com/g/cloudy-dev/c/69NLm1Ru_cs