Enable use of git tags in printed Cloudy versions
As part of the switch to git, a new workflow is to be established.
The previous practice of releasing every few years would better be
replaced by a scheme where we release every few months.  We seem to be
gravitating toward a model where feature branches are merged onto
master, which is merged onto a release branch for periodical releases.

The Cloudy version reported in the .out file treats the release branch
as any other, which would be opaque to users.  The version number stored
as a tag to the appropriate git commit is now employed as the Cloudy
version in user-facing runs.

Reported-by: M. Dehghanian
https://groups.google.com/g/cloudy-dev/c/N_tKJTriDDw