# additional entries for .gitignore
## The documentation

By default, we build both documentation and executables in folders
below master.  The code's documentation is in the
docs/latex and we generate PDF copies of the three parts
of Hazy and the QuickStart.  The default .gitignore does
ignore the latex helper files so only the final PDF needs be added.

## The executable

We build in the sys_* directories below source.  The default
.gitignore does ignore the compiler-generated files so only
a few implementation-dependent configuration files needed to be added.

## The test suite

The tsuite directory has several folders, a large number of models,
and an even larger number of files created during the test suite.
My workflow involves regularly testing with gcc and llvm and on
occasion with the Intel compiler.  All three versions of the
test suite should be available.  I created three symbolic link
folders, gtsuite, ltsuite, and itsuite, to hold the test suite
for the gcc, llvm, and icc compilers.  These three directories
are ignored.