Force consistency of equilibrium solves in non-eq sims
The command 'set dynamics population equilibrium' forced the solves for
the populations of neutrals and ions to their steady-state setup, but it
did not do so for molecules.  Add checks for the command's status in the
molecular network matrix setup to keep equilibrium treatment in
non-equilibrium runs consistent across the different branches of the code.

Reported-by: M. Chatzikos
https://groups.google.com/g/cloudy-dev/c/kFDQe9yq0Ts