Improve U and xi ionization parameters discussion
The notation in the description of the U and xi forms
of the ionization parameter was slightly inconsistent,
which caused problems for one user.  These are updated.
There is no change in content.

The TexShop app on macOS cares a number of helper files
to cross-reference the source and compiled text.  These
have been added to .gitignore