Gargi Shaw docs CR background and PAH interactions
This updates two parts of Hazy 1 and adds a citation to
Shaw & Ferland (2021arXiv210103732S).  The discussion
reviews how the PAH abundance affects the cosmic ray secondary
ionization rate and gives the recommended Galactic background
for the case where PAHs are present.  Our recommended default
CR background rate remains the Indriolo+ value but we give
the Cloudy command to change this rate to our recommended value
when PAHs are present.