Refactor dynamical iteration test into method
A plethora of syntaxes exists across the code for testing whether an
iteration is to be considered an initial relaxation iteration in dynamical
models.  For consistency, it is better for the task to be encapsulated in
a function, and to call that instead.

The only remaining instance of the check that is not replaced by the
method call occurs in source/conv_itercheck.cpp, which applies to
advective models.  The intent there is to guarantee that one iteration is
done after the initial ones with dynamical sources, and without declaring
convergence prematurely.

https://groups.google.com/g/cloudy-dev/c/3jd_sNqfCtE