merge changes in 4fe48fb5
This does an explicit merge of the changes that came onto master in 4fe48fb5.
The changes were a series of additions as branch maser and were in response
to thrown asserts in a grid of turbulent BLR sims.

c17_branch did not spawn off master so is in its own git universe and
these merges are necessary.