Update H-like n=2 energy to that of 2p (ignore 2s)
Taking the energy of 2s into account for the average energy of the n=2
configuration causes the H-like Lya wavelength to shift by a significant
amount.  In the case of Fe+25, the wavelength shifts by about 150 km/s,
comparable to turbulent broadening in Perseus.

Using the correct energy for the H-like n=2 configuration would limit
Cloudy's utility for X-ray microcalorimeter science.

NB NB The offset of He+ n=2 is maintained.