Fix bug in wl_checkall.pl
A bug fix on 2020-Aug-25 to fix_input_scripts_wl.pl was not fully ported
to wl_checkall.pl.  This caused some lines to be reported on screen as
unmatched, even though Cloudy proposed a match.

Remove duplication of subroutines defined in those scripts to prevent
that kind of error from occuring again.