Fix botches in all-double run
The update to NIST energies for the H and He iso-sequences triggered a
number of botches in the test run with FLT_IS_DBL.

One botch repeated in a few sims is a wavelength mis-identification of
a Balmer line near the Balmer jump, due to the higher precision of the
NIST energies.  The current precision of 6 significant digits in the
wavelength string is not sufficient: single precision runs round it to
3673.76A, double precision to .77A; while NIST quotes it as .7649A.
7-digit precision has been used for its specification, but such practice
is generally discouraged.

Other botches involved an Fe II band centered at 1212A, most likely
caused by slight changes to ionization potentials.  Their cause was not
investigated; tolerance bounds were widened to accept runs with both
precisions.