update tsuite to values with new UMIST & LAMDA
Update the test suite results to account for predictions
that changed as a result of Gargi Shaw updates to the chemistry
and internal structure of some molecules.

Several new LAMDA models were added which changes the temperare
of the gas, changing some line predictions, including
the CO rotation ladder in pdr_co_fully_noneq

The chemistry changes due to the UMIST updates and this should
affect molecular column densities. We monitor relatively
few molecular column densities.  Several changed by 0.3 dex.

The largest change was a line of HNC.  There were no direct
changes in the HNC model or chemistry.  This must be a trickle
down effect of changes to other molecules.