GS remove reactions driving v large changes in chemistry
These were new or revised reactions that appeared in
2012 UMIST but were not part of the network used in
the Leiden meeting.  They produced > 1 dex changes in
predicted column densities.  They are now commented out.

Gargi is researching these reactions to discover more
about their provenance.  It is possible that they
are an improvement but also that the changes were an
oversight or bug.