Fix bug with multiple 'save species lines' commands
Stefano Bianchi reported that for multiple instances of the command
'save species lines', only the first one is honored.

The code expects only one instance of this command, and relies on a
static bool 'lgRunOnce' to decide whether it has run.  As a result,
subsequent calls produce empty files.

Remove the bool, and replace it with a record of processed species
to guarantee that all requested species are processed, and each is
processed only once.

https://cloudyastrophysics.groups.io/g/Main/message/4612