Process FeII bands at all times
FeII emission being ubiquitous, it is good to always accumulate its
emission in the bands defined in the FeII_bands.ini file.  However,
with the previous setup, processing of the FeII bands was done only
when a 'save species bands' is issued.

Now the FeII emission is always summed up in these bands, so they can
be printed in the main output (.out).  If a 'save species bands'
command is given, the accumulated emission will be reported therein,
as well.