Fix bug with band emission accumulation
Commit 713ec4eac3031ead5d9dc4c902b9ea54f79a8443 attempted to fix a bug in
the way the emission in a band was accumulated.  The problem was that it
was unnecessarily reinitializing its emission vectors with every call, so
only the last zone was reported on output.

The fix was overzealous, and blindly adopted the logic used with the
accumulation of the pseudo-continuum, which, however, is not transferred,
unlike the band emission.  This led to the factor 'dVeffAper', which
expresses the effective volume modified by the aperture command, to be
applied twice.

The correct result is obtained by removing the extra factor.