Standardize 'save species bands' output intensities
The intensities produced by the command differ from those in the block
of lines in the main output.  The former intensities do not include the
corrective factor 'radius.Conv2PrtInten', and their values are not the
ones stored on the line stack.

Use of the same code for the command output as in the main output works
around these problems.