Forbid numbers on 'metals deplete' command
User Andreas Faisst reported that a scale factor issued with the
'metals deplete' command does not actually reset the abundances.
Cloudy is not designed to do both with a single command.  Hazy 1,
however, does not clearly state so.

Mitigate by explicitly forbidding numbers on the command, and by
updating Hazy 1 to clarify current policy.

https://cloudyastrophysics.groups.io/g/Main/message/4679