Fix bug in 'compile recombination coefficients H-like'
The command operates on either 40 or 20 shells, depending on the isoseq
and species. At the time of its inception, Cloudy computed all H-like
energies on the fly assuming hydrogenic scaling (1/n^2).  However, it
now uses NIST data, which typically do not reach n=20.

The current coding uses the number of resolved levels to distinguish
between NIST and hydrogenic energies, and leads to a negative energy and
a thrown ASSERT() when the command is used, because the expected data
are not in the relevant Stout energies file.

With this commit, hydrogenic energies are used when NIST energy data are
known, not expected, to be missing.