Add command 'print array level image'
The output of the command 'print array level' is difficult to handle
even under normal circumstances.  First because of the sheer number of
levels involved (for H by default, the matrix has dimensions 70x70),
and also because it is printed in the main output file, which can make
parsing the output challenging.

Create a visual representation of the matrix in the .ppm image format.

Derive the classes relevant to printing the rate matrices from the
module class, so as to be initialized automatically, without having to
explicitly invoke their zero() methods.