Let Chianti parser figure out number of CS points
Chianti version 7 allows for 5 or 9 spline points for the collision
strength of a transition.  By contrast, in version 10 a transition may
have as many as needed (e.g., 36 points for some transitions in O+2).

Prepare Cloudy for Chianti v10 datasets, allowing it to figure out on
the fly how many CS spline points a transition has.