Fix bug in Chianti10 Upsilon interpolation
An array of 9 doubles has been allocated on the stack for the
interpolation of collision strenghts, where 9 is the maximum number of
spline points used in Chianti v7.  This leads to stack smashing, and
program termination, when more spline points are available in Chianti
v10 data.

Adapt interpolation code to use the number of spline points of each
transition.