Make 'print arrays levels FITS' a 'save' command
Creating output files through the 'print' family of commands feels a
little awkward.  It is preferable to generate FITS or PPM images with
a subcommand of 'save', as done here.

Note, however, that this is not a proper 'save' command, since it
functions as a debugging aid.  It is therefore treated a little
differently than its counterparts both in parse_save.cpp, and in
save_do.cpp.