Fix indices for HFS, level2, inner shell transitions
These transitions are not associated with a particular atomic structure
model.  They are appended at the end of the transition stack, and are
given as level indices their position in the stack.  Confusingly, the
lower level indices are higher than the upper level indices.

Indices are swapped here, although a better treatment for may be
desired.