Fix bug in resolving n-n' collision strengths
Most n-n' collision strengths are currently over-resolved.  This affects
the CS of Vriens & Smeets (1980), Lebedev & Beigman (1998), and Fujimoto
(1978).  Percival and Richards (1978) are not subject to this bug.

The bug occurs because the statistical weights of the nl-resolved upper
and lower levels may be passed to these functions, but later their
result is further resolved by CSresolver().

The fix here omits passing the statistical weights of the nl-resolved
levels, but instead uses the statistical weights for n and n'; e.g.,
for H-like ions they are 2n^2.  nl-resolution of the total CS is done
by CSresolver(). All functions now operate the same as PR78.

The interface to functions VS80, LB98, and F78 CS are now consistent
with each other (iso-seq listed first, followed by n, n').

The code for Van Regemorter CS for H-like and He-like ions is now
refactored into a single function.