return to non-hyperfine split CN model
We updated the LAMDA model of CN to their hyperfine
split version.  Many CN lines that were multiplets
are now individual lines.  We now find one of the
lines and obtain its intensity rather than the full
intensity of the multiplet.  We would need to create
a number of create line blends to get the multiplet
intensity.  Those are inconvenient since then all lines
have the same label “Blnd” and so are hard to identify
in the output.

This change puts us back to the previous state with the
non-resolved model and the full multiplet intensity being
obtained.  The LAMDA masterlist file has both
versions present with a comment on how to go to
the fully hyperfine split model.

We need to look into the literature to see what is
actually observed.  Do most telescopes measure
the multiplet individual members?  What do most
observational studies do?