Fix bug with command order and 'save species bands'
Gary sent a bug report that triggered an error with setting up the
"FeII_bands.ini" file in a sim that contained a 'save' command, and
'init "honly.ini"', in that order.

The problem occurred because ParseSave() specified "FeII_bands.ini" for
loading.  Iron was disabled later, so the "Fe+" species was not entered
in the list of species known to the code.  This triggered an abort when
the code tried to set up the FeII bands.

Make sure that the species is active before attempting to set up bands.
This requires that the bands are initialized after the atomic data have
been read in -- the relevant calls have been moved to cloudy().

The initialization of pseudo continua has also been moved to cloudy(),
for symmetry.