Fix bug with FeII bands and consolidate classes
Hazy claims that the FeII bands are always processed, but this happens
only if a 'save' command has been given.  Move the command that enters
the FeII bands onto the stack of bands to be processed to the top of
SpeciesBandsCreate(), to make sure they are always added.

The 'save species bands' infrastructure uses two classes to store info
for each instance of the command.  The bands file and the species name
are stored as part of the t_save struct, while the data for summing the
band emission are stored in species_pseudo_cont.cpp.  Refactor the code
so that only one class exists that holds as data the parsed info as well
as an object of the class that sums up the emission.