Fix bug in statistical weight of singlet ground
The function that computes the statistical weight for levels in He-like
species reports 4n^2 for all levels.  The 4 sums over both singlets and
triplets.  However, the ground has only a singlet state, no triplet.

This is the corollary of an earlier bug, which made the same mistake in
how the n-n' CS is broken up among the fine structure levels of a He-like
ground state.