facilite doing Case B for H I, He I, and He II
Introduce variable to indicate which species we are doing.
Much of the remaining code can be derived from this variable.
Exceptions are the specification of the SED and size of the models.
So, largely, only the logical variables around lines 26-35 need be set.

The exceptions for particular ions occur around line 105