generate grid of He I / Case B predictions
This generates tables of current predictions
relative to Case B values, as set in the He I case B data
file in data directory.  This tests two very different
things.  Are the Case B predictions still valid or does that
file need to be updated?   Also, have our predictions changed
relative to that data file?

Both the data file and the predictions are volatile and we
cannot add this as a tsuite test asserting a ratio of unity.