Clarify maser string in output of 'save dr'
When zoning is due to an active maser, the above command reports, e.g.,
    H maser dTauMase=-2.00e-01 -10 7 3 0
as its cause.  This cryptic message is here refined to
    OH maser, levels 4 1, dTauMase=-2.00e-01
which explicitly declares the species name (previously 7), and the
meaning of the last two numbers ("3 0" -> "4 1") as referring to the
level indices in the (original, as of now) data file.

The constants (e.g., -10 above), which indicate the kind of lines each
transition is internally grouped with (e.g., UTA, HFS), are removed.