update the description of time-dependent commands
Some revisions are made in the section 14.4.1. They include adding more detailed explanation about how the time-table works. I also explained the situations in which the code stops its dynamical calculations. I added an example of a model with two radiation fields. Both radiation fields are lasers, while one of them is constant and the other one varies with time. I also explained that it is possible to vary both of the sources at the same time.