explain relationships between coarse & fine continua, and lines
There have been FAQs on the user forum in which
people expect to resolve the lines in the coarse continuum.
This add a discussion of the coarse and fine continua,
and explains that the lines are in their own space.
We recommend using SAVE LINE LIST to extract line intensities
and point to a second of Hazy 2 describing how to post process
our output to obtain line profiles.

 Changes to be committed:
	modified:   ../QuickStart/QuickStart.tex
	modified:   ../hazy1/conout.tex
	modified:   ../hazy1/cont-def.tex
	modified:   observed.tex