Update line pump rates and esc. prob at all times
To improve performance, Cloudy updates spectral line pump rates and
escape probabilities only once per zone, controlled with the flags
conv.lgFirstSweepThisZone and .lgLastSweepThisZone.

This design choice is old, and the performance gains may no longer be
as important as they once were.  In addition, it is known to prevent
masers from converging.

Remove the flags from the update logic to always carry it out when a
zone is processed.