add He I emissivity tests and improve doc of prog to compile emissivity table
Tests to generate ratios of our current Case B predictions relative to
the Porter+ predictions for a number of lines, densities, and temperatures.
We want the ratio to be unity.
	new file:   ../../experimental/he1n0t4.in
	new file:   ../../experimental/he1n2t4.in
	new file:   ../../experimental/he1n4t4.in
	new file:   ../../experimental/he1n6t4.in
	new file:   ../../experimental/he1ratio.dat

A table giving new/old Case B predictions for a number of lines,
densities, and temperatures
	new file:   he1nnt4.xlsx

Documentation update (in progress) for program
	new file:   readme.md
	deleted:    readme_vary_CaseBnt.htm