Update BibTeX acquisition in db-ref-bib2json.pl
The previous ADS BibTeX service (cgi-bin/nph-bib_query) no longer
exists.

As a result, storing which of the three databases (AST, PHY, and PRE)
is the origin of the BibTeX data has become obsolete.  Remove all
instances of use of the databases, and relevant data fields.

Instead, the new API requires users to supply a token in order
to query the server from the command-line.  Update to current method.

Use of the new API requires that an ADS token be used.  Users need
supply one only in interactive runs.