demote cautions of parts of SED being zero to notes
The demotes cautions that parts of the SED are zero to notes.
The following wc on auto shows that we have a large number of
these.  They are benign - they occur if we use a blackbody
that does go to zero far out in the Wien tail.  They are a
minor note about the SED rather than something deserving of
raising a caution.

the number of cautions in tsuite/auto went from

Alien20:auto:wc ~gary/cloudy/cautions.txt
  5267  72721 544951 /home/gary/cloudy/cautions.txt

to

Alien20:auto:wc cautions2.txt
  3846  64015 460146 cautions2.txt