Identify most opaque line in 'save fine continuum'
Anna Ogorzalek posted a query on the Cloudy group about identifying
absorption features in the output of the 'save xspec mtable' command.

A partial answer would be to identify features due to the lines alone.

The present adds two columns to the output of 'save fine continuum'.
The first identifies the species whose line is centered at that bin,
and the second reports the optical depth due that line alone (i.e., not
the total including any neighboring lines).