Report full line label in 'save fine continuum'
The previous implementation reported only a feature's species in
spectral notation, e.g., 'Fe17'.

This is useful if cells are not merged and the energy is reported in
wavelength units (say, Angstrom), because the exact wavelength is
reported in the first column.

Merging cells eliminates the wavelength information, and if the number
of cells merged together is high, the output can be confusing.

Report the entire spectral label, including the wavelength, to eliminate
ambiguity in the output.