Add comment to code block in LineStackCreate()
A (disabled) block of code was added at the end of LineStackCreate()
that worked its way through the map of fine continuum bins and line
stack indices to quantify the amount of overlap present.  The block
did not carry a comment.  During review, it was deemed necessary to
explain its intent.  This is done here.

Also, on the output of bins with overlap greater than 10 lines, the
line stack index is now reported to show that the repeating species
labels are indeed genuine.