Remove limit from code block in LineStackCreate()
The previous implementation of the code block in LineStackCreate() that
quantifies line overlap used an arbitrary value of 20 as the maximum
number of lines that could overlap in a bin.

During review, Gary questioned the choice.  This ad hoc value was picked
as a high enough limit to carry out the task, given that currently up to
14 lines overlap in a bin.

The previous implementation was also inefficient in that it went over
rfield.fine_lstack map several times to produce the report.

Here, the overlap across rfield.fine_lstack is established ahead of
time, and each fine continuum index in the map is visited only once for
the report.  The ad hoc value of 20 is removed.