Clarify header of 'save fine continuum'
Previously, the header produced by the command used the words 'Species'
and 'Opt Dep Spec'.  The former is outdated, given that we now report
the spectral line label (including wavelength), and the latter is not
clear to the user, given that we refer to the single-line optical depth
as Specific only in the source.

Here, the former is replaced with 'SpecLine' and the latter with
'Single Line Opt Depth'.