Enable the prototype GTH algorithm for all species
In the previous implementation, Robin had applied the prototype GTH
algorithm only for the pathological case of Ca 1.

With this changeset, the coverage is extended over all species (whose
sum of sources is 0).  It passes the test suite without any problems
and in fact it improves the wallclock performance of the code.

It is committed to allow for further testing with all the platforms we
support.  The hope is to include this GTH prototype in the upcoming
release.