Update fix_input_scripts_wl.pl for blends.ini
The file data/blends.ini and relevant infrastructure were introduced by
Peter van Hoof back in 2019 (2019-03-25 to 2019-04-09, according to the
SVN log).

However, the script fix_input_scripts_wl.pl was not updated to handle
updates to that file, partly because there had been no need for it.

The upgrade to Chianti v10 has required that the file be updated, and
the script is here modified to be able to carry such updates out.

It is also modified to handle PROBLEMs reported by Cloudy when, e.g., it
cannot identify a component to a line blend (a PROBLEM is followed by an
abort; updates must be done piecemeal).