Increase default number of levels for Fe 8 to 30
With the current setup, the test suite reports that it could not find
two Fe 8 lines at 130.941A and 270.090A used in agn_warm_absorber.in
and blr_n09_p22_Z20.in, respectively.

Inspection of the Chianti v7 fe_8.wgfa file reveals that these lines
correspond to transitions
  5 - 53   270.090   3s2.3p5.3d2 4D2.5 - 3s2.3p6.4f 2F2.5
  1 - 53   130.941   3s2.3p6.3d 2D1.5 - 3s2.3p6.4f 2F2.5

In Chianti v10, level 53 is indexed 64.

To recover these lines the default number of levels in CloudyChianti.ini
is here increased to 30 (the discrepancy being due to trimming by the
parser of levels with no experimental energy; the upper level index may
be found with the 'save lines labels' command).

Note that this recovers only the line at 130.941A.

The line at 270.090A is missing from the Chianti v10 data (fe_8.wgfa).