Fix Chianti transitions with AI & swapped levels
Chianti v10 contains transitions with the level indices swapped, meaning
that lower index > upper index.

Arrack checked only that the upper index be less than the maximum index
allowed for our AI cut.  In the case of fe_18, this let a few AI
transitions in the .wgfa file.

As a result, Cloudy crashed on a bounds-checking run.

Here, we remove fe_18.wgfa transitions with either index referring to AI
levels.  We also reverse the index order in some .wgfa and .splups files
so that the lower index appears first.