Initialize variables, as per g++ v12 warnings
Compilation with G++ v12.1.0 on a MacOS X (v12.4) has brought up a few
new warnings that relate mostly to initialization.

Most of these occur in functions of thirdparty_quadpack.cpp.  Initialize
the offending floats therein to NaN, and long integers to 0.

Also, the long members of the CollSplinesArray class are now initialized
to 0.