Fix syntax errors in sys_llvm/Makefile.conf
The previous implementation contained several errors, which were
revealed when compiling on older versions of MacOS.

First, the CXX variable used was the one defined in source/Makefile, not
the one defined (later) in Makefile.conf.  Hoist the definition to the
top of the file to make sure we use the right compiler when setting the
system architecture.

Second, there was a space before "shell", i.e., "$( shell" was used
instead of "$(shell", which caused the command to **not** run in the
shell, and prevented acquisition of the LLVM target.  The variable was
never set, the string always evaluated to the empty string, and the M1
CPU was always picked up.  This caused warnings on older versions of
MacOS X (Yosemite (10.10), and Monterey (12.0)).  Remove extra space.

Third, there was an extra space at the end of the conditional, which is
appended to the end of the result of findstring.  On M1/M2 systems, this
is innocuous because the strings don't match anyway.  But on Intel
systems, findstring evaluates to an empty string, but the right-hand
side of the conditional to a single space, which does not match the
empty string on the left-hand side.  This causes the CPU to be set to
the unsupported apple-m1.  Compilation aborts with an error.  Rearrange
the conditional to use quoted strings instead of parentheses -- this
makes the test robust against spaces.  Also, rearrange the conditional
to use 'ifeq' instead of 'ifneq', and set the architecture to the
'native' default in an else clause.