Add row options to 'save hyperfine populations'
Previously, when the command was used with the row keyword, it would
report, in one line of text, the ratio of level populations for all
spectral lines contained in the input file passed to the command.

Now the population ratio is still printed by default, but the individual
populations or the spin temperature may be reported, instead, by passing
an appropriate keyword.

This change is geared toward generalizing the command, and making more
useful, in preparation for merging it with the 'save line populations'
command.