Merge 'save populations' commands
Previously, there existed two commands for saving line populations:
'save line populations' reported the lower and upper level populations
for many (but not all) spectral lines known to the code, while
'save hyperfine populations' saved the populations and derived
quantities for a list of spectral lines (not necessarily hyperfine)
supplied by the user.

These are now brought together in a single scheme.  The command retains
its original name ('save line'), but the code it executes is that of the
new command ('save hyperfine').

The choice aims to simplify the output of the 'save line' command, which
produced large output files (several MB) even for single zone models.