Improve 'save line pop' logic
The previous implementation of the command rested on the assumption
that all line stack entries are associated with atomic models, chiefly
because it had been tested with real spectral lines.

However, a number of entries appear in the stack that are not real lines
and trying to access their populations leads to a segfault.

Define method LinSv.isSpectralLine() to tell if a line is a genuine
spectral line, defined here as either a transferred (type 't') or
recombination (type 'r') line.  Use method to screen against other lines
(e.g., blends) when producing the command output.

Also use a check on whether the line transition is associated, to screen
against recombination lines, which are not associated with an atomic
model either.

Note that some recombination lines are transferred, so a test against
line type 'r' is not sufficient.

Finally, protect the population ratio in a safe_div() to guard against
zero populations (e.g., when the relevant ion has zero density).