Simplify 'save line pop' check for lines with pop
Previously, a method was defined, LinSv.isSpectralLine(), that
discriminated against all line types and flags known to the code, except
for lines that are either transferred, 't', or are designated as
recombination, 'r', lines.  Then, when the method was needed there was
an additional check on the TransitionProxy of the line in the stack.

Robin wondered if all that was necessary, so upon a little reflection
it became apparent that they were not.  Instead of testing against all
those flags, it is more straightforward to test for the types that the
method does allow for: 't' and 'r'.

But given that only 't' lines with initialized transitions can ever pass
the if-statement in 'save lines pop', and produce meaningful output, the
method becomes superfluous.

It is hereby removed.  The if-statement now checks if the line is type
't' instead of calling that method to test if it is a genuine spectral
line.

The check against the transition being initialized (that discriminates
against recombination lines - which have no level populations associated
with them) remains.