Fix bug in run_programs.pl
The script instructs to run it with, say, sys_llvm.  However, when
queried about compilation flags, the Makefile in sys_* directories
returns 2 lines of output instead of the 1 line returned by the
Makefile in source.  This causes commands to be 4 lines instead of
1 for the sys_* directories, and the compilation to fail.

Capture only the last line of the Makefile output; compilation
commands are 1 line then, and the script works as intended.