Gari removes redundant reactions from mole_deuteriaum.dat
We had not tested master with D turned on. Even the primordial composition
set does not include D. Gargi ran the tsuite suite with D enabled and
found converge aborts and botched monitors on master. She carefully reviewed
the chemistry data and found that many D reactions were duplicated.
She fixed that and I am committing that one set of changes to
this branch.

I will propose merging it to master with D left off. The corrected
D chemistry will not be used by default, just as it is right now.
We will then merge from master with these changes to her HDgargi branch
and continue other HD development there.

There is no functional change in master as a result of this
set of changes, since D is not turned on by default.