Fix bug in gitversion.sh
When a release tag is cloned, the branch default to '(no branch)'.
The revision generated for the branch contains an unbalanced parenthesis
(i.e., it is '(no-c22.00'), which causes compilation to abort.

The present fix causes the revision to be, e.g., 'c22.00', on a cloned
release tag.

Recall, this appears in the first line of the main Cloudy output.