Fix bug in gitversion.sh
When a release tag is cloned, the branch default to '(no branch)'.
The revision generated for the branch contains an unbalanced parenthesis
(i.e., it is "(no-c22.00"), which causes compilation to abort.

The present fix causes the revision on a cloned release tag to be, e.g.,
'c22.00'.

Recall, this appears in the first line of the main Cloudy output.

NB NB This patch has been independently applied to the release branch.