Create temp fix with phfit.dat ion pot
It was reported that the increasing the continuum resolution by a factor
of 20, caused the code to fail with a 0 photionization cross section for
He-like Cu.

The problem was identified as a violation of DRY.  For some purposes
ionization potentials from NIST are used, while for others the older
values embedded in the file data/phfit.dat are used.  This mismatch
caused the thrown assert in the sim given to us.

While we are working on a permanent fix, we here update phfit.dat in the
release branch to guard against over-resolving the continuum by a factor
of 20.  Note this is not a permanent fix.  Using finer resolution causes
other ions to fail for similar reasons.

Reported-by: Tinggui Wang