Use NIST ion pot, instead of values in Stout
Previously, for H and He iso-sequence ions, the ionization potential was
obtained from the Stout energy files, ending in .nrg.dat.  A comparison
shows that these values are identical with the NIST values found in
ionization_potentials.dat.

Remove the IPs from the .nrg.dat files, and modify the data parser (in
iso.cpp) to use the NIST values instead.