Fix bug with rbegin() in mole_species.cpp
A reverse iterator is used to go over the list of nuclei present in a
molecule.  However, the iterator is decremented instead of incremented,
leading to a segmentation fault with LLVM on Mac.

Increment the iterator to fix the bug.  It now works with LLVM and GCC
on Mac, and GCC on Linux.

This is the only instance of a reverse_iterator being decremented in the
Cloudy codebase.  It occurs when the command 'no molecules heavy' is
issued (in fact, this is its only effect).

The bug has escaped detection because the command is not exercised in
the test suite.

Reported-by: Leoni Janssen, on https://cloudyastrophysics.groups.io/