Fix bug with H-like lines
There exists an asymmetry in the loops that computes the intensity of
H-like ions: while the loops goes over all the available levels, whether
a level is resolved or collapsed is decided based on local conditions,
which account for pressure lowering.

In a time-dependent simulation, this led to different lines being
entered in the stack.

The bug was discovered with a sim tailored to a problem Francis Keenan
brought attention to.

Reported-by: Robin Williams
Fixed-by: Robin Williams