Transfer H-like doublets fully
Previously, the H-like doublets had their populations set, but their
intensities were not computed.  Chamani Gunasekera did do that, and
these changes (in prt_lines.cpp) are included in this set.

Still, the line intensities are not reported on the output of the
'save emitted continuum' command.  The missing ingredient is adding
the lines to the rfield.outlin array.  This is achieved via use of the
outline() method of the ExtraLymanLinesJ*5 transitions.